+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-module-jackson`
+#### Added
+- Consider `@JsonProperty.name` override for methods
+- Look-up `"description"` for methods (if included) based on `@JsonProperty.description`
 
 ## [4.20.0] - 2021-09-04
 ### `jsonschema-generator`